Add futures.py
